stock: usa Decimal en schemas

+++ app/modules/stock/schema.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from pydantic import BaseModel, Field
 
 
@@ -8,19 +9,19 @@ class StockSucursalOut(BaseModel):
     producto_nombre: str
     nombre_variante: str
     sku: str | None = None
-    stock_fisico: float
-    stock_reservado: float
-    stock_vendido_pendiente_entrega: float
-    stock_disponible: float
+    stock_fisico: Decimal
+    stock_reservado: Decimal
+    stock_vendido_pendiente_entrega: Decimal
+    stock_disponible: Decimal
 
 
 class IngresoStockCreate(BaseModel):
     id_sucursal: int
     id_variante: int
     id_proveedor: int
-    cantidad_ingresada: float = Field(gt=0)
-    costo_productos: float = Field(ge=0)
-    gastos_adicionales: float = Field(ge=0, default=0)
+    cantidad_ingresada: Decimal = Field(gt=0)
+    costo_productos: Decimal = Field(ge=0)
+    gastos_adicionales: Decimal = Field(ge=0, default=0)
     origen_ingreso: str = "manual"
     observacion: str | None = None
     id_usuario: int
@@ -31,18 +32,18 @@ class IngresoStockResponse(BaseModel):
     ingreso_id: int
     id_sucursal: int
     id_variante: int
-    cantidad_ingresada: float
-    costo_total_lote: float
-    costo_unitario_calculado: float
-    stock_anterior: float
-    stock_nuevo: float
-    costo_promedio_anterior: float
-    costo_promedio_nuevo: float
+    cantidad_ingresada: Decimal
+    costo_total_lote: Decimal
+    costo_unitario_calculado: Decimal
+    stock_anterior: Decimal
+    stock_nuevo: Decimal
+    costo_promedio_anterior: Decimal
+    costo_promedio_nuevo: Decimal
 
 class AjusteStockCreate(BaseModel):
     id_sucursal: int
     id_variante: int
-    cantidad: float
+    cantidad: Decimal
     nota: str
     id_usuario: int
     origen_tipo: str = "ajuste_manual"
@@ -55,11 +56,11 @@ class AjusteStockResponse(BaseModel):
     id_sucursal: int
     id_variante: int
     tipo_movimiento: str
-    cantidad: float
-    stock_fisico_anterior: float
-    stock_reservado_anterior: float
-    stock_vendido_pendiente_entrega_anterior: float
-    stock_fisico_nuevo: float
-    stock_reservado_nuevo: float
-    stock_vendido_pendiente_entrega_nuevo: float
-    stock_disponible_nuevo: float
+    cantidad: Decimal
+    stock_fisico_anterior: Decimal
+    stock_reservado_anterior: Decimal
+    stock_vendido_pendiente_entrega_anterior: Decimal
+    stock_fisico_nuevo: Decimal
+    stock_reservado_nuevo: Decimal
+    stock_vendido_pendiente_entrega_nuevo: Decimal
+    stock_disponible_nuevo: Decimal
